Fix unsigned vs signed comparison warning in sigopcount_tests.cpp

+++ src/test/sigopcount_tests.cpp
@@ -20,21 +20,21 @@ BOOST_AUTO_TEST_CASE(GetSigOpCount)
 {
     // Test CScript::GetSigOpCount()
     CScript s1;
-    BOOST_CHECK_EQUAL(s1.GetSigOpCount(false), 0);
-    BOOST_CHECK_EQUAL(s1.GetSigOpCount(true), 0);
+    BOOST_CHECK_EQUAL(s1.GetSigOpCount(false), (unsigned int) 0);
+    BOOST_CHECK_EQUAL(s1.GetSigOpCount(true), (unsigned int) 0);
 
     uint160 dummy;
     s1 << OP_1 << dummy << dummy << OP_2 << OP_CHECKMULTISIG;
-    BOOST_CHECK_EQUAL(s1.GetSigOpCount(true), 2);
+    BOOST_CHECK_EQUAL(s1.GetSigOpCount(true), (unsigned int) 2);
     s1 << OP_IF << OP_CHECKSIG << OP_ENDIF;
-    BOOST_CHECK_EQUAL(s1.GetSigOpCount(true), 3);
-    BOOST_CHECK_EQUAL(s1.GetSigOpCount(false), 21);
+    BOOST_CHECK_EQUAL(s1.GetSigOpCount(true), (unsigned int) 3);
+    BOOST_CHECK_EQUAL(s1.GetSigOpCount(false), (unsigned int) 21);
 
     CScript p2sh;
     p2sh.SetDestination(s1.GetID());
     CScript scriptSig;
     scriptSig << OP_0 << Serialize(s1);
-    BOOST_CHECK_EQUAL(p2sh.GetSigOpCount(scriptSig), 3);
+    BOOST_CHECK_EQUAL(p2sh.GetSigOpCount(scriptSig), (unsigned int) 3);
 
     std::vector<CKey> keys;
     for (int i = 0; i < 3; i++)
@@ -45,15 +45,15 @@ BOOST_AUTO_TEST_CASE(GetSigOpCount)
     }
     CScript s2;
     s2.SetMultisig(1, keys);
-    BOOST_CHECK_EQUAL(s2.GetSigOpCount(true), 3);
-    BOOST_CHECK_EQUAL(s2.GetSigOpCount(false), 20);
+    BOOST_CHECK_EQUAL(s2.GetSigOpCount(true), (unsigned int) 3);
+    BOOST_CHECK_EQUAL(s2.GetSigOpCount(false), (unsigned int) 20);
 
     p2sh.SetDestination(s2.GetID());
-    BOOST_CHECK_EQUAL(p2sh.GetSigOpCount(true), 0);
-    BOOST_CHECK_EQUAL(p2sh.GetSigOpCount(false), 0);
+    BOOST_CHECK_EQUAL(p2sh.GetSigOpCount(true), (unsigned int) 0);
+    BOOST_CHECK_EQUAL(p2sh.GetSigOpCount(false), (unsigned int) 0);
     CScript scriptSig2;
     scriptSig2 << OP_1 << dummy << dummy << Serialize(s2);
-    BOOST_CHECK_EQUAL(p2sh.GetSigOpCount(scriptSig2), 3);
+    BOOST_CHECK_EQUAL(p2sh.GetSigOpCount(scriptSig2), (unsigned int) 3);
 }
 
 BOOST_AUTO_TEST_SUITE_END()
